+++ com.unity.render-pipelines.high-definition/Editor/Material/Fabric/ShaderGraph/FabricMasterNode.cs
@@ -797,5 +797,10 @@ namespace UnityEditor.Rendering.HighDefinition
 
             base.CollectShaderProperties(collector, generationMode);
         }
+
+        public override bool SupportsVirtualTexturing()
+        {
+            return true;
+        }
     }
 }