+++ com.unity.render-pipelines.high-definition/Editor/Material/Hair/ShaderGraph/HairMasterNode.cs
@@ -902,5 +902,10 @@ namespace UnityEditor.Rendering.HighDefinition
 
             base.CollectShaderProperties(collector, generationMode);
         }
+
+        public override bool SupportsVirtualTexturing()
+        {
+            return true;
+        }
     }
 }