+++ com.unity.shadergraph/Editor/Data/Interfaces/IGeneratesInclude.cs
@@ -1,0 +1,7 @@
+namespace UnityEditor.ShaderGraph
+{
+    interface IGeneratesInclude
+    {
+        void GenerateNodeInclude(IncludeRegistry registry, GenerationMode generationMode);
+    }
+}