+++ com.unity.shadergraph/Editor/Data/Nodes/IMasterNode.cs
@@ -8,6 +8,7 @@ namespace UnityEditor.ShaderGraph
     {
         string GetShader(GenerationMode mode, string name, out List<PropertyCollector.TextureInfo> configuredTextures, List<string> sourceAssetDependencyPaths = null);
         bool IsPipelineCompatible(RenderPipelineAsset renderPipelineAsset);
+        bool SupportsVirtualTexturing();
         int GetPreviewPassIndex();
         void ProcessPreviewMaterial(Material material);
     }