+++ com.unity.shadergraph/Editor/Data/Nodes/Input/Texture/TextureStackNode.cs
@@ -4,48 +4,94 @@ using UnityEditor.Graphing;
 using System.Collections.Generic;
 using System;
 using System.Globalization;
-using UnityEditor.ShaderGraph.Drawing.Controls;
+using UnityEditor.ShaderGraph.Drawing;
 using UnityEditor.ShaderGraph.Internal;
 using UnityEditor.Rendering;
 using UnityEngine.UIElements;
-using UnityEditor.ShaderGraph.Drawing;
 using UnityEditor.Graphing.Util;
+using UnityEngine.Rendering;
 
 namespace UnityEditor.ShaderGraph
 {
-    [Title("Input", "Texture", "Sample VT Stack")]
+    [Title("Input", "Texture", SampleTextureStackNode.DefaultNodeTitle)]
     [FormerName("UnityEditor.ShaderGraph.SampleTextureStackNodeBase")]
     [FormerName("UnityEditor.ShaderGraph.SampleTextureStackNode2")]
     [FormerName("UnityEditor.ShaderGraph.SampleTextureStackNode3")]
     [FormerName("UnityEditor.ShaderGraph.SampleTextureStackNode4")]
-    class SampleTextureStackNode : AbstractMaterialNode, IGeneratesBodyCode, IMayRequireMeshUV, IHasSettings
+    class SampleTextureStackNode : AbstractMaterialNode, IGeneratesBodyCode, IMayRequireMeshUV, IHasSettings, IGeneratesInclude, IMayRequireTime
     {
+        public const string DefaultNodeTitle = "Sample VT Stack";
+
         public const int UVInputId = 0;
         [NonSerialized]
         public readonly int[] OutputSlotIds = new int[] { 1, 2, 3, 4 };
         [NonSerialized]
-        public readonly int [] TextureInputIds = new int[] { 5, 6, 7, 8 };
+        public readonly int[] TextureInputIds = new int[] { 5, 6, 7, 8 };
         public const int FeedbackSlotId = 9;
         public const int LODInputId = 10;
+        public const int BiasInputId = 11;
+        public const int DxInputId = 12;
+        public const int DyInputId = 13;
 
         static string[] OutputSlotNames = { "Out", "Out2", "Out3", "Out4" };
         static string[] TextureInputNames = { "Texture", "Texture2", "Texture3", "Texture4" };
         const string UVInputNAme = "UV";
         const string FeedbackSlotName = "Feedback";
         const string LODSlotName = "Lod";
+        const string BiasSlotName = "Bias";
+        const string DxSlotName = "Dx";
+        const string DySlotName = "Dy";
 
         public override bool hasPreview { get { return false; } }
         bool isProcedural;// set internally only
 
+        // Keep these in sync with "VirtualTexturing.hlsl"
         public enum LodCalculation
         {
-            Automatic,
-            Explicit,
-            //Biased, //TODO: Add support to TextureStack.hlsl first
+            [InspectorName("Automatic")]
+            VtLevel_Automatic = 0,
+            [InspectorName("Lod Level")]
+            VtLevel_Lod = 1,
+            [InspectorName("Lod Bias")]
+            VtLevel_Bias = 2,
+            [InspectorName("Derivatives")]
+            VtLevel_Derivatives = 3
+        }
+
+        public enum AddresMode
+        {
+            [InspectorName("Wrap")]
+            VtAddressMode_Wrap = 0,
+            [InspectorName("Clamp")]
+            VtAddressMode_Clamp = 1,
+            [InspectorName("Udim")]
+            VtAddressMode_Udim = 2
+        }
+
+        public enum FilterMode
+        {
+            [InspectorName("Anisotropic")]
+            VtFilter_Anisotropic = 0
+        }
+
+        public enum UvSpace
+        {
+            [InspectorName("Regular")]
+            VtUvSpace_Regular = 0,
+            [InspectorName("Pre Transformed")]
+            VtUvSpace_PreTransformed = 1
+        }
+
+        public enum QualityMode
+        {
+            [InspectorName("Low")]
+            VtSampleQuality_Low = 0,
+            [InspectorName("High")]
+            VtSampleQuality_High = 1
         }
 
         [SerializeField]
-        LodCalculation m_LodCalculation = LodCalculation.Automatic;
+        LodCalculation m_LodCalculation = LodCalculation.VtLevel_Automatic;
         public LodCalculation lodCalculation
         {
             get
@@ -60,6 +106,24 @@ namespace UnityEditor.ShaderGraph
                 m_LodCalculation = value;
                 UpdateNodeAfterDeserialization();
                 Dirty(ModificationScope.Topological);
+            }
+        }
+
+        [SerializeField]
+        QualityMode m_SampleQuality = QualityMode.VtSampleQuality_High;
+        public QualityMode sampleQuality
+        {
+            get
+            {
+                return m_SampleQuality;
+            }
+            set
+            {
+                if (m_SampleQuality == value)
+                    return;
+
+                m_SampleQuality = value;
+                Dirty(ModificationScope.Node);
             }
         }
 
@@ -109,6 +173,26 @@ namespace UnityEditor.ShaderGraph
         }
 
         [SerializeField]
+        string m_StackName = "";
+
+        public string stackName
+        {
+            get
+            {
+                return m_StackName;
+            }
+            set
+            {
+                if (m_StackName == value)
+                    return;
+
+                m_StackName = value;
+                UpdateNodeAfterDeserialization();
+                Dirty(ModificationScope.Topological);
+            }
+        }
+
+        [SerializeField]
         protected TextureType[] m_TextureTypes = { TextureType.Default, TextureType.Default, TextureType.Default, TextureType.Default };
 
         // We have one normal/object space field for all layers for now, probably a nice compromise
@@ -129,6 +213,57 @@ namespace UnityEditor.ShaderGraph
             }
         }
 
+        protected string CleanupStackName(string name)
+        {
+            // Procedural stacks allow sharing the same name. This means they will sample the same data as well.
+            // This also allows sampling the same stack both from VS and PS.
+
+            if (isProcedural) return name;
+
+            // Make sure there is no other node with the same name
+            // if there is patch the current name by adding "_" so it's unique
+            var stacks = owner.GetNodes<SampleTextureStackNode>();
+            int tries = stacks.Count();
+
+            for (int i = 0; i < tries; i++)
+            {
+                bool conflict = false;
+                foreach (var node in stacks)
+                {
+                    if (node == this) continue;
+                    // Check if the names as emitted in the shader will clash
+                    if (node.GetStackName() == this.ProcessStackName(name) )
+                    {
+                        conflict = true;
+                        break;
+                    }
+                }
+                if (!conflict)
+                {
+                    return name;
+                }
+                // Try again to find a free one
+                name = name + "_";
+            }
+
+            return name;
+        }
+
+        /*
+            True if the masternode of the graph this node is currently in supports virtual texturing.
+        */ 
+        private bool supportedByMasterNode
+        {
+            get
+            {
+                var masterNode = owner?.GetNodes<IMasterNode>().FirstOrDefault();
+                return masterNode?.SupportsVirtualTexturing() ?? false;
+            }
+        }
+
+        /*
+            The panel behind the cogweel node settings
+        */
         class TextureStackNodeSettingsView : VisualElement
         {
             SampleTextureStackNode m_Node;
@@ -138,12 +273,54 @@ namespace UnityEditor.ShaderGraph
 
                 PropertySheet ps = new PropertySheet();
 
+                ps.Add(new PropertyRow(new Label("Stack Name")), (row) =>
+                {
+                    row.Add(new IdentifierField(), (field) =>
+                    {
+                        field.value = m_Node.stackName;
+                        field.isDelayed = true;
+                        field.RegisterValueChangedCallback(evt =>
+                        {
+                            var clean = m_Node.CleanupStackName(evt.newValue);
+                            if (m_Node.stackName == clean)
+                                return;
+
+                            m_Node.owner.owner.RegisterCompleteObjectUndo("Stack Name Change");
+                            m_Node.stackName = (string)clean;
+                            field.SetValueWithoutNotify(m_Node.stackName);// Make sure the cleaned up name is used
+                        });
+                    });
+                });
+
                 ps.Add(new PropertyRow(new Label("Lod Mode")), (row) =>
                 {
                     row.Add(new UIElements.EnumField(m_Node.lodCalculation), (field) =>
                     {
                         field.value = m_Node.lodCalculation;
-                        field.RegisterValueChangedCallback(ChangeLod);
+                        field.RegisterValueChangedCallback(evt =>
+                        {
+                            if (m_Node.lodCalculation == (LodCalculation)evt.newValue)
+                                return;
+
+                            m_Node.owner.owner.RegisterCompleteObjectUndo("Lod Mode Change");
+                            m_Node.lodCalculation = (LodCalculation)evt.newValue;
+                        });
+                    });
+                });
+
+                ps.Add(new PropertyRow(new Label("Quality")), (row) =>
+                {
+                    row.Add(new UIElements.EnumField(m_Node.sampleQuality), (field) =>
+                    {
+                        field.value = m_Node.sampleQuality;
+                        field.RegisterValueChangedCallback(evt =>
+                        {
+                            if (m_Node.sampleQuality == (QualityMode)evt.newValue)
+                                return;
+
+                            m_Node.owner.owner.RegisterCompleteObjectUndo("Quality Change");
+                            m_Node.sampleQuality = (QualityMode)evt.newValue;
+                        });
                     });
                 });
 
@@ -152,28 +329,52 @@ namespace UnityEditor.ShaderGraph
                     row.Add(new UIElements.IntegerField(), (field) =>
                     {
                         field.value = m_Node.numSlots;
-                        field.RegisterValueChangedCallback(ChangeNumSlots);
+                        field.isDelayed = true;
+                        field.RegisterValueChangedCallback(evt =>
+                        {
+                            if (Equals(m_Node.numSlots, evt.newValue))
+                                return;
+
+                            m_Node.owner.owner.RegisterCompleteObjectUndo("NumSlots Flow Change");
+                            m_Node.numSlots = evt.newValue;
+                            field.SetValueWithoutNotify(m_Node.numSlots);//This may be clamped
+                        });
                     });
                 });
 
-                ps.Add(new PropertyRow(new Label("No Resolve")), (row) =>
+                ps.Add(new PropertyRow(new Label("No Feedback")), (row) =>
                 {
                     row.Add(new UnityEngine.UIElements.Toggle(), (field) =>
                     {
                         field.value = m_Node.noFeedback;
-                        field.RegisterValueChangedCallback(ChangeFeedback);
+                        field.RegisterValueChangedCallback(evt =>
+                        {
+                            if (m_Node.noFeedback == evt.newValue)
+                                return;
+
+                            m_Node.owner.owner.RegisterCompleteObjectUndo("Feedback Settings Change");
+                            m_Node.noFeedback = evt.newValue;
+                        });
                     });
                 });
 
-                for (int i=0; i<node.numSlots; i++)
+                for (int i = 0; i < node.numSlots; i++)
                 {
                     int currentIndex = i; //to make lambda by-ref capturing happy
-                    ps.Add(new PropertyRow(new Label("Type " + i)), (row) =>
+                    ps.Add(new PropertyRow(new Label("Layer " + (i + 1) + " Type")), (row) =>
                     {
                         row.Add(new UIElements.EnumField(m_Node.m_TextureTypes[i]), (field) =>
                         {
                             field.value = m_Node.m_TextureTypes[i];
-                            field.RegisterValueChangedCallback(evt => { ChangeTextureType(evt, currentIndex); } );
+                            field.RegisterValueChangedCallback(evt =>
+                            {
+                                if (m_Node.m_TextureTypes[currentIndex] == (TextureType)evt.newValue)
+                                    return;
+
+                                m_Node.owner.owner.RegisterCompleteObjectUndo("Texture Type Change");
+                                m_Node.m_TextureTypes[currentIndex] = (TextureType)evt.newValue;
+                                m_Node.Dirty(ModificationScope.Graph);
+                            });
                         });
                     });
                 }
@@ -183,56 +384,32 @@ namespace UnityEditor.ShaderGraph
                     row.Add(new UIElements.EnumField(m_Node.normalMapSpace), (field) =>
                     {
                         field.value = m_Node.normalMapSpace;
-                        field.RegisterValueChangedCallback(ChangeNormalMapSpace);
+                        field.RegisterValueChangedCallback(evt =>
+                        {
+                            if (m_Node.normalMapSpace == (NormalMapSpace)evt.newValue)
+                                return;
+
+                            m_Node.owner.owner.RegisterCompleteObjectUndo("Normal Map space Change");
+                            m_Node.normalMapSpace = (NormalMapSpace)evt.newValue;
+                        });
                     });
                 });
 
+#if !ENABLE_VIRTUALTEXTURES
+                ps.Add(new HelpBoxRow(MessageType.Warning), (row) => row.Add(new Label("VT is disabled, this node will do regular 2D sampling.")));
+#endif
+                if (!m_Node.supportedByMasterNode)
+                {
+                    ps.Add(new HelpBoxRow(MessageType.Warning), (row) => row.Add(new Label("The current master node does not support VT, this node will do regular 2D sampling.")));
+                }
+
+                IVirtualTexturingEnabledRenderPipeline vtRp = GraphicsSettings.currentRenderPipeline as IVirtualTexturingEnabledRenderPipeline;
+                if (vtRp == null || vtRp.virtualTexturingEnabled == false)
+                {
+                    ps.Add(new HelpBoxRow(MessageType.Warning), (row) => row.Add(new Label("The current render pipeline does not support VT." + ((vtRp ==null) ? "(Interface not implemented by" + GraphicsSettings.currentRenderPipeline.GetType().Name + ")" : "(virtualTexturingEnabled == false)"))));
+                }
+
                 Add(ps);
-            }
-
-            void ChangeNumSlots(ChangeEvent<int> evt)
-            {
-                if (Equals(m_Node.numSlots, evt.newValue))
-                    return;
-
-                m_Node.owner.owner.RegisterCompleteObjectUndo("NumSlots Flow Change");
-                m_Node.numSlots = evt.newValue;
-            }
-
-            void ChangeLod(ChangeEvent<Enum> evt)
-            {
-                if (m_Node.lodCalculation == (LodCalculation)evt.newValue)
-                    return;
-
-                m_Node.owner.owner.RegisterCompleteObjectUndo("Lod Mode Change");
-                m_Node.lodCalculation = (LodCalculation)evt.newValue;
-            }
-
-            void ChangeTextureType(ChangeEvent<Enum> evt, int index)
-            {
-                if (m_Node.m_TextureTypes[index] == (TextureType)evt.newValue)
-                    return;
-
-                m_Node.owner.owner.RegisterCompleteObjectUndo("Texture Type Change");
-                m_Node.m_TextureTypes[index] = (TextureType)evt.newValue;
-            }
-
-            void ChangeFeedback(ChangeEvent<bool> evt)
-            {
-                if (m_Node.noFeedback == evt.newValue)
-                    return;
-
-                m_Node.owner.owner.RegisterCompleteObjectUndo("Feedback Settings Change");
-                m_Node.noFeedback = evt.newValue;
-            }
-
-            void ChangeNormalMapSpace(ChangeEvent<Enum> evt)
-            {
-                if (m_Node.normalMapSpace == (NormalMapSpace)evt.newValue)
-                    return;
-
-                m_Node.owner.owner.RegisterCompleteObjectUndo("Normal Map space Change");
-                m_Node.normalMapSpace = (NormalMapSpace)evt.newValue;
             }
         }
 
@@ -241,7 +418,7 @@ namespace UnityEditor.ShaderGraph
             return new TextureStackNodeSettingsView(this);
         }
 
-        public SampleTextureStackNode() : this(1) {}
+        public SampleTextureStackNode() : this(1) { }
 
         public SampleTextureStackNode(int numSlots, bool procedural = false, bool isLod = false, bool noResolve = false)
         {
@@ -273,19 +450,13 @@ namespace UnityEditor.ShaderGraph
             {
                 usedSlots.Add(FeedbackSlotId);
             }
-            if (m_LodCalculation != LodCalculation.Automatic)
-            {
-                usedSlots.Add(LODInputId);
-            }
-
-            usedSlots.ToArray();
 
             // Create slots
             AddSlot(new UVMaterialSlot(UVInputId, UVInputNAme, UVInputNAme, UVChannel.UV0));
 
             for (int i = 0; i < numSlots; i++)
             {
-                AddSlot(new Vector4MaterialSlot(OutputSlotIds[i], OutputSlotNames[i], OutputSlotNames[i], SlotType.Output, Vector4.zero, (noFeedback && m_LodCalculation == LodCalculation.Explicit) ? ShaderStageCapability.All : ShaderStageCapability.Fragment));
+                AddSlot(new Vector4MaterialSlot(OutputSlotIds[i], OutputSlotNames[i], OutputSlotNames[i], SlotType.Output, Vector4.zero, (noFeedback && m_LodCalculation == LodCalculation.VtLevel_Lod) ? ShaderStageCapability.All : ShaderStageCapability.Fragment));
             }
 
             if (!isProcedural)
@@ -296,10 +467,28 @@ namespace UnityEditor.ShaderGraph
                 }
             }
 
-            if (m_LodCalculation != LodCalculation.Automatic)
+            if (m_LodCalculation == LodCalculation.VtLevel_Lod)
             {
                 var slot = new Vector1MaterialSlot(LODInputId, LODSlotName, LODSlotName, SlotType.Input, 0.0f, ShaderStageCapability.All, LODSlotName);
+                usedSlots.Add(LODInputId);
                 AddSlot(slot);
+            }
+
+            if (m_LodCalculation == LodCalculation.VtLevel_Bias)
+            {
+                var slot = new Vector1MaterialSlot(BiasInputId, BiasSlotName, BiasSlotName, SlotType.Input, 0.0f, ShaderStageCapability.Fragment, BiasSlotName);
+                usedSlots.Add(BiasInputId);
+                AddSlot(slot);
+            }
+
+            if (m_LodCalculation == LodCalculation.VtLevel_Derivatives)
+            {
+                var slot1 = new Vector2MaterialSlot(DxInputId, DxSlotName, DxSlotName, SlotType.Input, Vector2.one, ShaderStageCapability.All, DxSlotName);
+                var slot2 = new Vector2MaterialSlot(DyInputId, DySlotName, DySlotName, SlotType.Input, Vector2.one, ShaderStageCapability.All, DySlotName);
+                usedSlots.Add(DxInputId);
+                usedSlots.Add(DyInputId);
+                AddSlot(slot1);
+                AddSlot(slot2);
             }
 
             if (!noFeedback)
@@ -310,19 +499,63 @@ namespace UnityEditor.ShaderGraph
             }
 
             RemoveSlotsNameNotMatching(usedSlots, true);
+            name = String.IsNullOrEmpty(m_StackName) ? DefaultNodeTitle : m_StackName;
+        }
+
+        public static bool SubGraphHasStacks(SubGraphNode node)
+        {
+            var asset = node.asset;
+            foreach (var input in asset.inputs)
+            {
+                var texInput = input as Texture2DShaderProperty;
+                if (texInput != null && !string.IsNullOrEmpty(texInput.textureStack))
+                {
+                    return true;
+                }
+            }
+            return false;
+        }
+
+        public static List<string> GetSubGraphInputStacks(SubGraphNode node)
+        {
+            // There could be more stacks in the subgraph but as they are not part of inputs they don't
+            // leak out so we don't case about those.
+            List<string> result = new List<string>(); //todo pooled list
+            var asset = node.asset;
+            foreach (var input in asset.inputs)
+            {
+                var texInput = input as Texture2DShaderProperty;
+                if (texInput != null && !string.IsNullOrEmpty(texInput.textureStack))
+                {
+                    result.Add(texInput.textureStack);
+                }
+            }
+            return result;
         }
 
         public static void ValidatNodes(GraphData d)
         {
-            ValidatNodes(d.GetNodes<SampleTextureStackNode>());
+            ValidatNodes(d.GetNodes<SampleTextureStackNode>(), d.GetNodes<SubGraphNode>());
         }
 
-        public static void ValidatNodes(IEnumerable<SampleTextureStackNode> nodes)
+        public static void ValidatNodes(IEnumerable<SampleTextureStackNode> nodes, IEnumerable<SubGraphNode> subNodes)
         {
-            List<KeyValuePair<string, string>> slotNames = new List<KeyValuePair<string, string>>();
+            var valueNameLookup = new Dictionary<string, string>();
+            var nodeNames = new HashSet<string>();
 
             foreach (SampleTextureStackNode node in nodes)
             {
+                if (nodeNames.Contains(node.GetStackName()) && !node.isProcedural)
+                {
+                    // Add a validation error, values need to be unique
+                    node.owner.AddValidationError(node.tempId, $"Some stack nodes have the same name '{node.GetStackName()}', please ensure all stack nodes have unique names.", ShaderCompilerMessageSeverity.Error);
+                }
+                else
+                {
+                    // Procedural nodes are still added here as we disallow procedural and regular nodes with the same name...
+                    nodeNames.Add(node.GetStackName());
+                }
+
                 for (int i = 0; i < node.numSlots; i++)
                 {
                     if (!node.isProcedural)
@@ -331,40 +564,52 @@ namespace UnityEditor.ShaderGraph
                         string name = node.FindSlot<MaterialSlot>(node.TextureInputIds[i]).displayName;
 
                         // Check if there is already a slot with the same value
-                        int found = slotNames.FindIndex(elem => elem.Key == value);
-                        if (found >= 0)
+                        string displayName;
+                        if (valueNameLookup.TryGetValue(value, out displayName))
                         {
                             // Add a validation error, values need to be unique
-                            node.owner.AddValidationError(node.tempId, $"Slot stack input slot '{value}' shares it's input with another stack input '{slotNames[found].Value}'. Please make sure every slot has unique input textures attached to it.", ShaderCompilerMessageSeverity.Error);
+                            node.owner.AddValidationError(node.tempId, $"Input slot '{name}' shares it's value '{value}' with another stack '{displayName}'. Please make sure every slot has unique input textures attached to it.", ShaderCompilerMessageSeverity.Error);
                         }
                         else
                         {
                             // Save it for checking against other slots
-                            slotNames.Add(new KeyValuePair<string, string>(value, name));
+                            valueNameLookup.Add(value, node.GetStackName() + " (slot " + name + ")");
                         }
                     }
+                }
+            }
 
-#if PROCEDURAL_VT_IN_GRAPH
-                    // Check if there is already a node with the same sampleid
-                    SampleTextureStackProceduralNode ssp = node as SampleTextureStackProceduralNode;
-                    if (ssp != null)
+            foreach (SubGraphNode node in subNodes)
+            {
+                var subStacks = GetSubGraphInputStacks(node);
+                foreach (var subStack in subStacks)
+                {
+                    // Todo how to exclude procedurals in subgraphs?
+                    if (nodeNames.Contains(subStack))
                     {
-                        string value = ssp.GetStackName();
-                        string name = ssp.GetStackName();
-                        // Check if there is already a slot with the same value
-                        int found = slotNames.FindIndex(elem => elem.Key == value);
-                        if (found >= 0)
-                        {
-                            // Add a validation error, values need to be unique
-                            node.owner.AddValidationError(node.tempId, $"This node has the same procedural ID as another node. Nodes need to have different procedural IDs.", ShaderCompilerMessageSeverity.Error);
-                        }
-                        else
-                        {
-                            // Save it for checking against other slots
-                            slotNames.Add(new KeyValuePair<string, string>(value, name));
-                        }
+                        node.owner.AddValidationError(node.tempId, $"A stack node in a subgraph which is exposed through a texture argument has the same name '{subStack}', please ensure all stack nodes have unique names across the whole shader using them.", ShaderCompilerMessageSeverity.Error);
                     }
-#endif
+                    else
+                    {
+                        nodeNames.Add(subStack);
+                    }
+                }
+
+                Dictionary<string, string> valueToStackLookup = node.GetValueToTextureStackDictionary(GenerationMode.ForReals);
+                foreach (var kvp in valueToStackLookup)
+                {
+                    // Check if there is already a slot with the same value
+                    string stackName;
+                    if (valueNameLookup.TryGetValue(kvp.Key, out stackName))
+                    {
+                        // Add a validation error, values need to be unique
+                        node.owner.AddValidationError(node.tempId, $"Stack '{kvp.Value}' shares it's value '{kvp.Key}' with another stack '{stackName}'. Please make sure every slot has unique input textures attached to it.", ShaderCompilerMessageSeverity.Error);
+                    }
+                    else
+                    {
+                        // Save it for checking against other slots
+                        valueNameLookup.Add(kvp.Key, kvp.Value);
+                    }
                 }
             }
         }
@@ -389,10 +634,23 @@ namespace UnityEditor.ShaderGraph
             get { return PreviewMode.Preview3D; }
         }
 
+        protected virtual string ProcessStackName(string potentialName)
+        {
+            // We do a poor man's namespacing here since the user's entered name could clash with existing symbols defined in the shaders or templates
+            // adding variables in templates later could also suddenly break user's stacks so any stacks are prefixed by a "namespace".
+            if (!string.IsNullOrEmpty(potentialName))
+            {
+                return "StackNodeNamespace_" + potentialName;
+            }
+            else
+            {
+                return string.Format("TexStack_{0}", GuidEncoder.Encode(guid));
+            }
+        }
 
         protected virtual string GetStackName()
         {
-            return GetVariableNameForSlot(OutputSlotIds[0]) + "_texturestack";
+            return ProcessStackName(m_StackName);
         }
 
         private string GetTextureName(int layerIndex, GenerationMode generationMode)
@@ -407,39 +665,56 @@ namespace UnityEditor.ShaderGraph
             }
         }
 
-        private string GetSampleFunction()
+        string MakeVtParameters(string variableName, string uvExpr, string lodExpr, string dxExpr, string dyExpr, AddresMode address, FilterMode filter, LodCalculation lod, UvSpace space, QualityMode quality)
         {
-            if (m_LodCalculation != LodCalculation.Automatic)
-            {
-                return "SampleStackLod";
-            }
-            else
-            {
-                return "SampleStack";
-            }
+            const string VTParametersInputTemplate = @"
+                        VtInputParameters {0};
+                        {0}.uv = {1};
+                        {0}.lodOrOffset = {2};
+                        {0}.dx = {3};
+                        {0}.dy = {4};
+                        {0}.addressMode = {5};
+                        {0}.filterMode = {6};
+                        {0}.levelMode = {7};
+                        {0}.uvMode = {8};
+                        {0}.sampleQuality = {9};
+            ";
+
+            return string.Format(VTParametersInputTemplate,
+                variableName,
+                uvExpr,
+                (string.IsNullOrEmpty(lodExpr)) ? "0.0f" : lodExpr,
+                (string.IsNullOrEmpty(dxExpr)) ? "float2(0.0f, 0.0f)" : dxExpr,
+                (string.IsNullOrEmpty(dyExpr)) ? "float2(0.0f, 0.0f)" : dyExpr,
+                address.ToString(),
+                filter.ToString(),
+                lod.ToString(),
+                space.ToString(),
+                quality.ToString());
+        }
+
+        string MakeVtSample(string infoVariable, string textureName, string outputVariableName, LodCalculation lod, QualityMode quality)
+        {
+            const string SampleTemplate = @"$precision4 {0} = SampleStack({1}, {2}, {3}, {4});";
+
+            return string.Format(SampleTemplate,
+                outputVariableName,
+                infoVariable,
+                lod.ToString(),
+                quality.ToString(),
+                textureName);
         }
 
         // Node generations
         public virtual void GenerateNodeCode(ShaderStringBuilder sb, GenerationMode generationMode)
         {
-            // This is not in templates or headers so this error only gets checked in shaders actually using the VT node
-            // as vt headers get included even if there are no vt nodes yet.
-            sb.AppendLine("#if defined(_SURFACE_TYPE_TRANSPARENT)");
-            sb.AppendLine("#error VT cannot be used on transparent surfaces.");
-            sb.AppendLine("#endif");
-            sb.AppendLine("#if defined(SHADERPASS) && (SHADERPASS == SHADERPASS_DBUFFER_PROJECTOR)"); //SHADERPASS is not defined for preview materials so check this first.
-            sb.AppendLine("#error VT cannot be used on decals. (DBuffer)");
-            sb.AppendLine("#endif");
-            sb.AppendLine("#if defined(SHADERPASS) && (SHADERPASS == SHADERPASS_DBUFFER_MESH)");
-            sb.AppendLine("#error VT cannot be used on decals. (Mesh)");
-            sb.AppendLine("#endif");
-            sb.AppendLine("#if defined(SHADERPASS) && (SHADERPASS == SHADERPASS_FORWARD_EMISSIVE_PROJECTOR)");
-            sb.AppendLine("#error VT cannot be used on decals. (Projector)");
-            sb.AppendLine("#endif");
-
             // Not all outputs may be connected (well one is or we wouldn't get called) so we are careful to
             // only generate code for connected outputs
+
             string stackName = GetStackName();
+            string localVariablePrefix = GetVariableNameForNode();
+            string parametersVariableNme = localVariablePrefix + "_pars";
+            string infoVariableName = localVariablePrefix + "_info";
 
             bool anyConnected = false;
             for (int i = 0; i < numSlots; i++)
@@ -455,34 +730,36 @@ namespace UnityEditor.ShaderGraph
 
             if (anyConnected)
             {
-                if (m_LodCalculation == LodCalculation.Automatic)
-                {
-                    string result = string.Format("StackInfo {0}_info = PrepareStack({1}, {0});"
-                        , stackName
-                        , GetSlotValue(UVInputId, generationMode));
-                    sb.AppendLine(result);
-                }
-                else
-                {
-                    string result = string.Format("StackInfo {0}_info = PrepareStackLod({1}, {0}, {2});"
-                        , stackName
-                        , GetSlotValue(UVInputId, generationMode)
-                        , GetSlotValue(LODInputId, generationMode));
-                    sb.AppendLine(result);
-                }
+                sb.Append(MakeVtParameters(
+                    parametersVariableNme,
+                    GetSlotValue(UVInputId, generationMode),
+                    (lodCalculation == LodCalculation.VtLevel_Lod) ? GetSlotValue(LODInputId, generationMode) : GetSlotValue(BiasInputId, generationMode),
+                    GetSlotValue(DxInputId, generationMode),
+                    GetSlotValue(DyInputId, generationMode),
+                    AddresMode.VtAddressMode_Wrap,
+                    FilterMode.VtFilter_Anisotropic,
+                    m_LodCalculation,
+                    UvSpace.VtUvSpace_Regular,
+                    m_SampleQuality));
+
+                sb.AppendLine(string.Format("StackInfo {0} = PrepareStack({1}, {2});"
+                                        , infoVariableName
+                                        , parametersVariableNme
+                                        , stackName));
             }
 
             for (int i = 0; i < numSlots; i++)
             {
                 if (IsSlotConnected(OutputSlotIds[i]))
                 {
-                    var id = GetTextureName(i, generationMode);
-                    string resultLayer = string.Format("$precision4 {1} = {3}({0}_info, {2});"
+                    var textureName = GetTextureName(i, generationMode);
+                    /*string resultLayer = string.Format("$precision4 {1} = {3}({0}_info, {2});"
                             , stackName
                             , GetVariableNameForSlot(OutputSlotIds[i])
-                            , id
+                            , textureName
                             , GetSampleFunction());
-                    sb.AppendLine(resultLayer);
+                    sb.AppendLine(resultLayer);*/
+                    sb.AppendLine(MakeVtSample(infoVariableName, textureName, GetVariableNameForSlot(OutputSlotIds[i]), m_LodCalculation, m_SampleQuality));
                 }
             }
 
@@ -508,11 +785,40 @@ namespace UnityEditor.ShaderGraph
             if (!noFeedback && feedbackConnected)
             {
                 //TODO: Investigate if the feedback pass can use halfs
-                string feedBackCode = string.Format("float4 {0} = GetResolveOutput({1}_info);",
+                string feedBackCode = string.Format("float4 {0} = GetResolveOutput({1});",
                         GetVariableNameForSlot(FeedbackSlotId),
-                        stackName);
+                        infoVariableName);
                 sb.AppendLine(feedBackCode);
             }
+        }
+
+        public void GenerateNodeInclude(IncludeRegistry registry, GenerationMode generationMode)
+        {
+            // This is not in templates or headers so this error only gets checked in shaders actually using the VT node
+            // as vt headers get included even if there are no vt nodes yet.
+            IVirtualTexturingEnabledRenderPipeline vtRp = GraphicsSettings.currentRenderPipeline as IVirtualTexturingEnabledRenderPipeline;
+
+            if (!supportedByMasterNode)
+            {
+                // The master node is not white listed for VT just use regular textures even if vt is on for this project.
+                registry.ProvideIncludeBlock("disable-vt-if-unsupported-node", "#define FORCE_VIRTUAL_TEXTURING_OFF 1");
+            }
+            else if (vtRp == null || vtRp.virtualTexturingEnabled == false)
+            {
+                // The master render pipeline does not support VT just use regular textures even if vt is on for this project.
+                registry.ProvideIncludeBlock("disable-vt-if-unsupported-node", "#define FORCE_VIRTUAL_TEXTURING_OFF 1");
+            }
+            else
+            {
+                // The master node supports VT but certain keyword settings in child materials may still break it. We catch this here.
+                registry.ProvideIncludeBlock("disable-vt-if-unsupported", @"#if defined(_SURFACE_TYPE_TRANSPARENT)
+#define FORCE_VIRTUAL_TEXTURING_OFF 1
+#error VT cannot be used on transparent surfaces.
+#endif");
+            }
+
+            // Always include the header even if vt is off this ensures the macros providing fallbacks are existing.
+            registry.ProvideInclude("Packages/com.unity.render-pipelines.core/ShaderLibrary/TextureStack.hlsl");
         }
 
         public override void CollectShaderProperties(PropertyCollector properties, GenerationMode generationMode)
@@ -579,6 +885,11 @@ namespace UnityEditor.ShaderGraph
                 return false;
             }
         }
+
+        public bool RequiresTime()
+        {
+            return true; //HACK: This ensures we repaint in shadergraph so data that gets streamed in also becomes visible.
+        }
     }
 
     class TextureStackAggregateFeedbackNode : AbstractMaterialNode, IGeneratesBodyCode, IMayRequireRequirePixelCoordinate
@@ -620,7 +931,7 @@ namespace UnityEditor.ShaderGraph
 
             if (numSlots == 1)
             {
-                string feedBackCode = $"float4 {GetVariableNameForSlot(AggregateOutputId)} = {GetSlotValue(AggregateInputFirstId, generationMode)};";
+                string feedBackCode = $"float4 {GetVariableNameForSlot(AggregateOutputId)} = GetPackedVTFeedback({GetSlotValue(AggregateInputFirstId, generationMode)});";
                 sb.AppendLine(feedBackCode);
             }
             else if (numSlots > 1)
@@ -636,7 +947,7 @@ namespace UnityEditor.ShaderGraph
                     arrayIndex++;
                 }
 
-                string feedBackCode = $"float4 {GetVariableNameForSlot(AggregateOutputId)} = {arrayName}[ (IN.{ShaderGeneratorNames.PixelCoordinate}.x  + _FrameCount )% (uint){numSlots}];";
+                string feedBackCode = $"float4 {GetVariableNameForSlot(AggregateOutputId)} = GetPackedVTFeedback({arrayName}[ (IN.{ShaderGeneratorNames.PixelCoordinate}.x  + _FrameCount )% (uint){numSlots}]);";
 
                 sb.AppendLine(feedBackCode);
             }
@@ -654,15 +965,21 @@ namespace UnityEditor.ShaderGraph
     static class VirtualTexturingFeedback
     {
         public const int OutputSlotID = 22021982;
+        public const string subgraphOutputFrefix = "VTFeedbackIn_";
 
         // Automatically add a  streaming feedback node and correctly connect it to stack samples are connected to it and it is connected to the master node output
         public static IMasterNode AutoInject(IMasterNode iMasterNode)
         {
+            Debug.Log("Inject vt feedback");
             var masterNode = iMasterNode as AbstractMaterialNode;
             var stackNodes = GraphUtil.FindDownStreamNodesOfType<SampleTextureStackNode>(masterNode);
+            var subgraphNodes = GraphUtil.FindDownStreamNodesOfType<SubGraphNode>(masterNode);
+            Debug.Log("SubgraphNodes: " + subgraphNodes.Count);
 
-            // Early out if there are no VT nodes in the graph
-            if (stackNodes.Count <= 0)
+
+            // Try to find out early if there are no nodes doing VT feedback in the graph
+            // this avoids unnecessary work of copying the graph and patching it.
+            if (stackNodes.Count <= 0 && subgraphNodes.Count <= 0)
             {
                 return iMasterNode;
             }
@@ -670,13 +987,28 @@ namespace UnityEditor.ShaderGraph
             bool hasFeedback = false;
             foreach (var node in stackNodes)
             {
-                if ( !node.noFeedback )
+                if (!node.noFeedback)
                 {
                     hasFeedback = true;
+                    break;
                 }
             }
+
+            if (!hasFeedback) foreach (var node in subgraphNodes)
+                {
+                    foreach (var slot in node.GetOutputSlots<Vector4MaterialSlot>())
+                    {
+                        if (slot.shaderOutputName.StartsWith(subgraphOutputFrefix))
+                        {
+                            hasFeedback = true;
+                            break;
+                        }
+                    }
+                }
+
             if (!hasFeedback)
             {
+                Debug.Log("No vt feedback early out");
                 return iMasterNode;
             }
 
@@ -684,7 +1016,7 @@ namespace UnityEditor.ShaderGraph
             var workingMasterNode = masterNode.owner.ScratchCopy().GetNodeFromGuid(masterNode.guid);
 
             // inject VTFeedback output slot
-            var vtFeedbackSlot = new Vector4MaterialSlot(OutputSlotID, "VTFeedback", "VTFeedback", SlotType.Input, Vector4.one, ShaderStageCapability.Fragment);
+            var vtFeedbackSlot = new Vector4MaterialSlot(OutputSlotID, "VTPackedFeedback", "VTPackedFeedback", SlotType.Input, Vector4.one, ShaderStageCapability.Fragment);
             vtFeedbackSlot.hidden = true;
             workingMasterNode.AddSlot(vtFeedbackSlot);
 
@@ -714,6 +1046,23 @@ namespace UnityEditor.ShaderGraph
                 i++;
             }
 
+            foreach (var node in subgraphNodes)
+            {
+                foreach (var slot in node.GetOutputSlots<Vector4MaterialSlot>())
+                {
+                    if (!slot.shaderOutputName.StartsWith(subgraphOutputFrefix)) continue;
+
+                    // Create a new slot on the aggregate that is similar to the uv input slot
+                    string name = "FeedIn_" + i;
+                    var newSlot = new Vector4MaterialSlot(TextureStackAggregateFeedbackNode.AggregateInputFirstId + i, name, name, SlotType.Input, Vector4.zero, ShaderStageCapability.Fragment);
+                    newSlot.owner = feedbackNode;
+                    feedbackNode.AddSlot(newSlot);
+
+                    feedbackNode.owner.Connect(slot.slotReference, newSlot.slotReference);
+                    i++;
+                }
+            }
+
             // Add input to master node
             var feedbackInputSlot = workingMasterNode.FindInputSlot<ISlot>(OutputSlotID);
             if (feedbackInputSlot == null)
@@ -734,6 +1083,66 @@ namespace UnityEditor.ShaderGraph
 
             return workingMasterNode as IMasterNode;
         }
+
+        // Automatically add a  streaming feedback node and correctly connect it to stack samples are connected to it and it is connected to the master node output
+        public static SubGraphOutputNode AutoInjectSubgraph(SubGraphOutputNode masterNode)
+        {
+            var stackNodes = GraphUtil.FindDownStreamNodesOfType<SampleTextureStackNode>(masterNode);
+
+            // Early out if there are no VT nodes in the graph
+            if (stackNodes.Count <= 0)
+            {
+                Debug.Log("No vt in subgr");
+                return masterNode;
+            }
+
+            bool hasFeedback = false;
+            foreach (var node in stackNodes)
+            {
+                if (!node.noFeedback)
+                {
+                    hasFeedback = true;
+                }
+            }
+            if (!hasFeedback)
+            {
+                Debug.Log("No feedback in subgr");
+                return masterNode;
+            }
+
+            // Duplicate the Graph so we can modify it
+            Debug.Log("clonign graph");
+            var workingMasterNode = masterNode.owner.ScratchCopy().GetNodeFromGuid(masterNode.guid);
+
+            // Add inputs to feedback node
+            int i = 0;
+            foreach (var node in stackNodes)
+            {
+                // Find feedback output slot on the vt node
+                var stackFeedbackOutputSlot = (node.FindOutputSlot<ISlot>(SampleTextureStackNode.FeedbackSlotId)) as Vector4MaterialSlot;
+                if (stackFeedbackOutputSlot == null)
+                {
+                    // Nodes which are noResolve don't have a resolve slot so just skip them
+                    Debug.Log("No feedback slot on node, skipping");
+                    continue;
+                }
+
+                // Create a new slot on the master node for each vt feedback
+                string name = subgraphOutputFrefix + i;
+                var newSlot = new Vector4MaterialSlot(OutputSlotID + i, name, name, SlotType.Input, Vector4.zero, ShaderStageCapability.Fragment);
+                newSlot.owner = workingMasterNode;
+                newSlot.hidden = true;
+                workingMasterNode.AddSlot(newSlot);
+
+                workingMasterNode.owner.Connect(stackFeedbackOutputSlot.slotReference, newSlot.slotReference);
+                i++;
+
+                Debug.Log("Added feedback " + name);
+            }
+
+            workingMasterNode.owner.ClearChanges();
+            return workingMasterNode as SubGraphOutputNode;
+        }
     }
 
 #if PROCEDURAL_VT_IN_GRAPH
@@ -744,30 +1153,6 @@ namespace UnityEditor.ShaderGraph
     {
         public SampleTextureStackProceduralNode() : base(1, true)
         { }
-
-        [IntegerControl("Sample ID")]
-        public int sampleID
-        {
-            get { return m_sampleId; }
-            set
-            {
-                if (m_sampleId == value)
-                    return;
-
-                m_sampleId = value;
-                Dirty(ModificationScope.Graph);
-
-                ValidateNode();
-            }
-        }
-
-        [SerializeField]
-        int m_sampleId = 0;
-
-        protected override string GetStackName()
-        {
-            return "Procedural" + m_sampleId;
-        }
     }
 #endif
 }