+++ com.unity.shadergraph/Editor/Data/SubGraph/SubGraphAsset.cs
@@ -42,6 +42,8 @@ namespace UnityEditor.ShaderGraph
 
         public List<FunctionPair> functions = new List<FunctionPair>();
 
+        public List<FunctionPair> includes = new List<FunctionPair>();
+
         [NonSerialized]
         public List<AbstractShaderProperty> inputs = new List<AbstractShaderProperty>();